refactor: rename metadata package to schema
Package previously named "metadata" only contained schema-related types
(SchemaService, SchemaUpdateResult, and their tests). Renaming to "schema"
makes the package name accurate to its contents.

Moves preserve git history via git mv. Imports updated in Main, MainTest,
and McpToolRegistrationTest. Spec and plan paths updated.
