+++ src/test/java/org/apache/solr/mcp/server/McpToolRegistrationTest.java
@@ -24,7 +24,7 @@ import java.util.Arrays;
 import java.util.List;
 import org.apache.solr.mcp.server.collection.CollectionService;
 import org.apache.solr.mcp.server.indexing.IndexingService;
-import org.apache.solr.mcp.server.metadata.SchemaService;
+import org.apache.solr.mcp.server.schema.SchemaService;
 import org.apache.solr.mcp.server.search.SearchService;
 import org.junit.jupiter.api.Test;
 import org.springaicommunity.mcp.annotation.McpTool;
