+++ src/test/java/org/apache/solr/mcp/server/schema/SchemaServiceIntegrationTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.solr.mcp.server.metadata;
+package org.apache.solr.mcp.server.schema;
 
 import static org.junit.jupiter.api.Assertions.*;
